feat: breaking PostList in Container
